Simplify rendering
